Refactored code to separate some of the functionality to other files

+++ src/te-builder.js
@@ -1,10 +1,11 @@
 #!/usr/bin/env node
-import temp from 'temp';
 import fs from 'fs-extra';
 import path from 'path';
 import inquirer from 'inquirer-promise';
 import os from 'os';
-import archiver from 'archiver';
+import temp from 'temp';
+import { replaceFilenames } from './file/fileUtils.js';
+import { zipDirectory } from './zip/zipBuilder.js';
 
 const REPLACE_FILE_MAP = {
   "TypeExtensionD1": "$TypeExtensionD1"
@@ -15,15 +16,6 @@ const customerQuestion = {
   name: 'customer',
   message: 'Please enter the customer shorthand: '
 };
-
-//TODO Check if current directory contains a PlatforModule.xml. If it does, just build this.
-
-//If it doesn't, then ask the user to enter the customer shorthand
-//Check if the customer folder exists. If it doesn't, exit with an error message
-//If it does, then give a selector prompt for which folder to build
-//Once the user selects it, copy all the contents to a temp folder
-//Rename the TypeExtensionD1 to $TypeExtensionD1
-//Zip the contents of the temp folder to a current folder using the original folder name as the zip filename, and zip the contents to the cwd
 
 function getFolderList(pathToCheck) {
   const items = fs.readdirSync(pathToCheck);
@@ -36,6 +28,11 @@ function getFolderList(pathToCheck) {
   return folders;
 }
 
+/**
+ * Queries the user to find the location of the folder that houses the platform module
+ * 
+ * @returns string
+ */
 function questionPlatformModuleFolder() {
   const defaultPlatformHome = path.join(os.homedir(), 'code', 'gtnexus', 'platform');
   let customer, platformModuleFolder;
@@ -65,77 +62,18 @@ function questionPlatformModuleFolder() {
 }
 
 /**
- * Replace all files from the depth first search based on the REPLACE_FILE_MAP
+ * Copies files in this folder to a temporary folder, renames the appropriate files based on REPLACE_FILE_MAP, and then zips it to the current working directory.
  * 
- * @param {string} folderPath 
+ * @param {string} platformModuleFolder 
  */
-function replaceFilenames(folderPath) {
-  const stack = [folderPath]; // Initialize stack with the root folder
-
-  while (stack.length > 0) {
-    const currentPath = stack.pop(); // Pop the path from the end of the stack
-
-    const contents = fs.readdirSync(currentPath);
-
-    contents.forEach(item => {
-      const itemPath = path.join(currentPath, item);
-      const stat = fs.lstatSync(itemPath);
-
-      if (stat.isDirectory()) {
-        // If it's a directory, add it to the stack for processing
-        stack.push(itemPath);
-      } else {
-        // If it's a file, check if it needs to be renamed
-        const filename = path.basename(itemPath);
-        if (REPLACE_FILE_MAP.hasOwnProperty(filename)) {
-          const parentFolder = path.dirname(itemPath);
-          const newName = REPLACE_FILE_MAP[filename];
-          const newPath = path.join(parentFolder, newName);
-          fs.renameSync(itemPath, newPath);
-          console.log(`Renamed file: ${itemPath} -> ${newPath}`);
-        }
-      }
-    });
-
-    // Rename folder if it is in REPLACE_FILE_MAP;
-    const folderBasename = path.basename(currentPath);
-    if (currentPath !== folderPath && REPLACE_FILE_MAP[folderBasename]) {
-      const parentFolder = path.dirname(currentPath);
-      const newName = REPLACE_FILE_MAP[folderBasename];
-      const newPath = path.join(parentFolder, newName);
-      fs.renameSync(currentPath, newPath);
-      console.log(`Renamed folder: ${currentPath} -> ${newPath}`);
-    }
-  }
-}
-
-function zipDirectory(sourceDir, targetFile) {
-  return new Promise((resolve, reject) => {
-    const output = fs.createWriteStream(targetFile);
-    const archive = archiver('zip', { zlib: { level: 9 } });
-
-    output.on('close', function () {
-      resolve(targetFile);
-    });
-
-    archive.on('error', function (err) {
-      reject(err);
-    });
-
-    archive.pipe(output);
-    archive.directory(sourceDir, false);
-    archive.finalize();
-  });
-}
-
-function createZip(platformModuleFolder) {
+function createPlatformModuleZip(platformModuleFolder) {
   let tempDir;
   const currentDir = process.cwd();
   try {
     const platformModule = path.basename(platformModuleFolder);
     tempDir = temp.mkdirSync(platformModule);
     fs.copySync(platformModuleFolder, tempDir);
-    replaceFilenames(tempDir);
+    replaceFilenames(tempDir, REPLACE_FILE_MAP);
     const zipFilename = path.join(currentDir, `${platformModule}.zip`);
   
     zipDirectory(tempDir, zipFilename).then(() => {
@@ -144,7 +82,7 @@ function createZip(platformModuleFolder) {
     });
   } catch(err) {
     console.error(`Caught error while creating zip.\n${err.stack}`);
-    if (tempDir) {
+    if (tempDir && fs.existsSync(tempDir)) {
       fs.removeSync(tempDir);
     }
   }
@@ -156,12 +94,12 @@ function main() {
 
   if (fs.existsSync(platformModuleXmlPath)) {
     console.log('PlatformModule.xml found in the current directory. Building...');
-    createZip(currentDir);
+    createPlatformModuleZip(currentDir);
   } else {
     questionPlatformModuleFolder()
       .then((platformModuleFolder) => {
         console.log(`platformModuleFolder = ${platformModuleFolder}`);
-        createZip(platformModuleFolder);
+        createPlatformModuleZip(platformModuleFolder);
       }).catch(err => {
         console.error('Extraction failed:', err);
         process.exit(1);
